feat(condo): DOMA-8513 Extract common parts of import wrappers into BaseImportWrapper.tsx
